Added dot function to frontend paddle API
There is a slight difference that I am not sure how to handle, the dot function in paddle can take a 2-D array which will give an output to each corresponding pair in x and y arrays, then return a list with a n values equal to the number of arrays inside the 2-D array. I am not sure if it needs to be implemented though.